refactor(logging): one logging mechanism — migrate all 90 raw fprintf sites in src/
The engine had two ways to emit: IMP_LOG_* (level-gated, timestamped, file:line)
at ~1300 sites, and raw fprintf(stderr, ...) at 191. The second obeyed nothing.
#1221 gave the framework a working level for the first time, which made the
split matter: set diagnostics.log_level=error and you still got [gemm-algo],
[DEBUG_FWD] and [DEBUG_TPL] on stderr, untimestamped and unsuppressable.

All 90 sites in src/ now go through the framework. tools/ keeps its fprintf —
a CLI writing to stderr is program output, not logging, and that line is where
the layering already is.

Two things worth recording rather than hiding:

- MY OWN CONVERSION SCRIPT INTRODUCED A BUG. fprintf(stderr, "\n") is a line
  terminator, not a message; the mechanical pass turned ten of them into
  IMP_LOG_DEBUG("") — empty log lines. It compiled clean and every test stayed
  green, because nothing asserts on debug output. It surfaced only because a
  later exact-text match failed. Those blocks are now assembled properly:
  build the line, log it once.
- log_message() had no printf format checking. Added
  __attribute__((format(printf, 4, 5))), which is what verifies a migration of
  this shape: zero mismatches in the 90 conversions, and five PRE-EXISTING ones
  in weight_upload.cu (%lld against int64_t, which is long on LP64) that had
  been invisible. Fixed here.

Verified at runtime, not just compiled: with log_level=debug the engine emits
360 DEBUG lines, zero empty ones, and the hand-assembled blocks read correctly —
"input_tokens (16): 1 1 1 ...", "positions (16): 0 1 2 ...", and the stat line
with its optional NaN/Inf suffix.

NOT runtime-verified: the chatml/llama3/phi/harmony token dumps and the decode
dump in chat_template_families.cpp (this model takes the Jinja path), the
[PPL-DUMP] block and the [LOGITS] top8 block. They reuse the same two patterns
that are verified above, but they were not executed.

ctest -L unit 4/4, clang-format clean on changed lines.

+++ src/model/weight_upload.cu
@@ -1171,7 +1171,8 @@ static bool upload_layer_attention_weights(TransformerLayer& L, int i, const Upl
                 IMP_LOG_ERROR("Failed to dequant GPTQ %s for layer %d", name, i);
                 return false;
             }
-            IMP_LOG_DEBUG("GPTQ dequant %s layer %d -> [%lld, %lld] FP16", name, i, w.shape[0], w.shape[1]);
+            IMP_LOG_DEBUG("GPTQ dequant %s layer %d -> [%lld, %lld] FP16", name, i, (long long)w.shape[0],
+                          (long long)w.shape[1]);
         }
     }
 
@@ -1267,7 +1268,8 @@ static bool upload_layer_attention_weights(TransformerLayer& L, int i, const Upl
     // rope_freqs: upload as raw FP32 (NOT converted to FP16).
     // The RoPE kernel reads these as float* — FP16 conversion would corrupt them.
     if (L.rope_freqs.data && !L.rope_freqs.on_device && L.rope_freqs.qtype == QType::F32) {
-        IMP_LOG_INFO("Layer %d: uploading rope_freqs as raw FP32 (%lld elements)", i, L.rope_freqs.numel());
+        IMP_LOG_INFO("Layer %d: uploading rope_freqs as raw FP32 (%lld elements)", i,
+                     (long long)L.rope_freqs.numel());
         size_t bytes = L.rope_freqs.nbytes();
         void* d_data = nullptr;
         IMP_CUDA_CHECK_LOG(cudaMallocAsync(&d_data, bytes, ctx.stream));
@@ -1311,7 +1313,8 @@ static bool upload_layer_ffn_weights(TransformerLayer& L, int i, const UploadCtx
                 IMP_LOG_ERROR("Failed to dequant GPTQ %s for layer %d", name, i);
                 return false;
             }
-            IMP_LOG_DEBUG("GPTQ dequant %s layer %d -> [%lld, %lld] FP16", name, i, w.shape[0], w.shape[1]);
+            IMP_LOG_DEBUG("GPTQ dequant %s layer %d -> [%lld, %lld] FP16", name, i, (long long)w.shape[0],
+                          (long long)w.shape[1]);
         }
     }
